Location + Tracking
Converted the doubles to floats and made an approximation script to calculate how far the GPS cordinates are off and how much tollerance there is.

+++ Wonder Route/Assets/Scripts/Locations/Locations.cs
@@ -9,11 +9,11 @@ public class Locations : MonoBehaviour
     private Location locationData;
 
     //public enum Location {SintLucasIngang, Microlab, PSVStadium, MCdonalds, End};
-    double[,] gpsData; // 
+    float[,] gpsData; // 
 
     void Start()
     {
-        gpsData = new double[4, 2];
+        gpsData = new float[4, 2];
 
         SetLocation();
     }
@@ -27,12 +27,12 @@ public class Locations : MonoBehaviour
             gpsData[i, 1] = i + 2;
             Debug.Log(gpsData[i, 1]);
         }
-        gpsData[0, 0] = 51.44761661871976;
+        gpsData[0, 0] = 51.44761661871976f;
     }
 
-    public double[] GetLocation(Location loc)
+    public float[] GetLocation(Location loc)
     {
-        double[] gpsCords = new double[gpsData.GetLength(1)];
+        float[] gpsCords = new float[gpsData.GetLength(1)];
         int locInt = (int)loc;
 
         for (int i = 0; i < gpsCords.Length; i++)
